normalise tracked URLs to contain trailing slashes
The approach is conservative; it only applies if the URL has a protocol, and
no query fields or fragments. If the last part of the URL segment contains
a dot it is assumed to be a file and left alone. So, if it is such a URL,
and does not end with a slash, a slash is added. This is applied to
Action and Label immediately before sending, so URLs are always
normalised in this way, however they are generated.